services/flowise-proxy-service-py/requirements.txt
fastapi==0.104.1
hypercorn==0.14.4
pyjwt==2.8.0
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic-settings==2.0.3
pydantic[email]
flowise==1.0.4
uvicorn[standard]
json-repair==0.47.6
pillow==10.0.1
python-magic==0.4.27